DEV add else
Make result of last execution to be result of execution of script. Result of script now is single String, not list of Strings.
Rename empty to void.
Rename keyword to be BLOCKWORD.
Add else BLOCKWORD.
Add elif BLOCKWORD
Execute else in while and if blocwords if condition failed.
Add ElseExecutionToken to always return true in condition and act as replacement for "elif (true)"

Signed-off-by: Stanislav <56927835+Stanislavius@users.noreply.github.com>